requirements.txt
# Core Langchain and LangGraph dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langgraph>=0.1.5
langgraph-checkpoint-mongodb>=0.0.5
# langgraph-store

# Language model integrations
langchain-google-genai>=0.0.5
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# Persistence and database
pymongo>=4.6.0
motor>=3.3.0

# Web Framework and API
fastapi>=0.105.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Audio processing and speech recognition
deepgram-sdk>=2.11.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.3
scipy>=1.11.3

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
pydantic-settings>=2.0.0
fastapi-oauth2>=0.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.1
slowapi>=0.1.9
radon
pylint
pyaudio



# Secure Code Execution
docker>=7.0.0

google-genai>=0.0.0
langgraph.store.mongodb
radon
pylint
pyaudio
pydantic[email]
passlib
python-jose
itsdangerous
fastapi_oauth2